Text File IO
Re-added file IO, now using a text-based system.

Currently relies of a chain of type-checking if statements, which is less than ideal. Will likely change this in the future.

Also separated the add/edit charge UI from the AddChargeListener class. This will be used to edit charges in the future.