Arrow Alpha
Arrows can now be rendered with a variable alpha value.

Currently, arrow alpha is determined by the electrical field strength at a position. ArrowView uses a cubic falloff based on the maximum arrow magnitude present to determine the alpha value of individual arrows. In the future, I would like change this falloff calculation to not be dependent on the position of the arrow with the strongest field strength.